requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.4
django-phonenumber-field==6.1.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
phonenumbers==8.12.48
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
vonage==2.6.4
whitenoise==6.1.0
wrapt==1.14.1
